web/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
redis>=5
pydantic>=2
python-dotenv>=1
